модуляция версия 1.0
