+++ js/setup-const.js
@@ -1,0 +1,50 @@
+const PICTURE_COUNT = 25;
+const AVATAR_COUNT = 6;
+const LIKE_MIN_COUNT = 15;
+const LIKE_MAX_COUNT = 200;
+const MAX_COMMENT_INDEX = 1000;
+const MAX_COMMENT_COUNT = 5;
+const COMMENT_MESSEGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const DESCRIPTIONS = [
+  'Описание фото1', 'Описание фото2', 'Описание фото3', 'Описание фото4', 'Описание фото5', 'Описание фото6', 'Описание фото7',
+  'Описание фото8', 'Описание фото9', 'Описание фото10', 'Описание фото11', 'Описание фото12', 'Описание фото13', 'Описание фото14',
+  'Описание фото15', 'Описание фото16', 'Описание фото17', 'Описание фото18', 'Описание фото19', 'Описание фото20'
+];
+
+const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+  'Мартин',
+  'Педро',
+  'Рамзес',
+  'Пауло',
+  'Пабло',
+  'Митрофан',
+  'Иван Степанович',
+  'Волан-де-Морт',
+  'Азазелло',
+  'Пендальф Серый'
+];
+
+export { PICTURE_COUNT };
+export { AVATAR_COUNT };
+export { LIKE_MIN_COUNT };
+export { LIKE_MAX_COUNT };
+export { MAX_COMMENT_INDEX };
+export { MAX_COMMENT_COUNT };
+export { COMMENT_MESSEGE };
+export { DESCRIPTIONS };
+export { NAMES };
